requirements.txt
fastapi~=0.124.0
SQLAlchemy~=2.0.44
pydantic-settings~=2.12.0
passlib~=1.7.4
python-jose~=3.5.0
pydantic~=2.12.5
starlette~=0.50.0
python-dotenv~=1.2.1
protobuf~=6.33.2
alembic~=1.17.2